+++ src/css/main.css
@@ -30,7 +30,6 @@ body,
   overflow: hidden;
   position: absolute;
   top: 0;
-  left: 0;
 }
 
 .stack-top {
