+++ src/index.html
@@ -6,7 +6,7 @@
     <link rel="icon" type="image/png" sizes="32x32" href="img/favicon_io/favicon-32x32.png" />
     <link rel="icon" type="image/png" sizes="16x16" href="img/favicon_io/favicon-16x16.png" />
     <link rel="manifest" href="img/favicon_io/site.webmanifest" />
-    <title>[terrestris] 3D Demo Client</title>
+    <title></title>
   </head>
 
   <body>
